requirements.txt
requests
python-dotenv
openai
fastapi
uvicorn
beautifulsoup4